requirements.txt
# Web Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration & Environment
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Metrics
prometheus-client==0.19.0

# Utilities
psutil==5.9.6
python-dateutil==2.8.2

# Async
aiofiles==23.2.1

# Logging
python-json-logger==2.0.7

# AI / Ollama
openai>=1.0.0